# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SnipVault Sync Server API
  version: 3.0.1
  summary: HTTP API for the self-hosted SnipVault sync server.
  description: |
    The **SnipVault sync server** is the SnipVault web app run as a container. It
    exposes a small REST API that the desktop app uses to keep one prompt/snippet
    library in sync across several machines, plus health/status probes for
    operators.

    ## Authentication

    Every `/api/**` route is guarded by a bearer token, checked in constant time
    by the server's middleware (`proxy.ts`). Send it as
    `Authorization: Bearer <token>`.

    The server runs in one of three auth modes, chosen by environment:

    | Mode | Env | Behaviour |
    | --- | --- | --- |
    | **Single-user** | `SNIPVAULT_TOKEN` | One token guards one shared library (`data/snippets.db`). |
    | **Multi-user** | `SNIPVAULT_TOKENS=alice:tokA,bob:tokB` | Each user's token maps to a **private vault** (`data/users/<user>/snippets.db`). Structurally isolated — one user cannot read another's data. If both env vars are set, `SNIPVAULT_TOKENS` wins. |
    | **No-auth** | neither set | Requests pass through in development; in production the server **fails closed** (`503`) unless `SNIPVAULT_ALLOW_NO_AUTH=1`. |

    In multi-user mode the middleware resolves the caller's user from their token
    and forwards it internally on a trusted `x-snipvault-user` header; any inbound
    value of that header from a client is always stripped, so it cannot be
...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/sync:
    post:
      tags: [Sync]
      operationId: sync
      summary: Bidirectional sync
      description: |
        Reconcile a client's full record set with the server's in one round trip:

        1. The client POSTs its full record set (including its own tombstones).
        2. The server merges each record by `uuid`, newest `updated_at` winning.
        3. The server returns its full, now-merged set so the client can apply the
           same newest-wins rule locally.

        Both sides converge to the union of records with the most recent edits.
        Unknown languages are preserved (a newer client may have languages this
        server doesn't know yet).

        **Limits:** request body ≤ 25 MB; ≤ 100000 records per request; per-snippet
        code ≤ 500000 chars and description ≤ 100000 chars. Exceeding a size/count
        cap returns `413`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                records:
                  type: array
                  description: The client's full record set. A missing/non-array value is treated as empty.
                  items:
                    $ref: "#/components/schemas/SyncRecord"
      responses:
        "200":
          description: The server's full, merged record set plus how many were applied.
          content:
            application/json:
              schema:
                type: object
                required: [records, applied]
                properties:
                  records:
                    type: array
                    items:
                      $ref: "#/components/schemas/SyncRecord"
                  applied:
                    type: integer
                    description: Number of incoming records inserted or updated on the server.
                    example: 12
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          description: Payload too large — body over 25 MB or more than 100000 records.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/ServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Per-user (multi-user mode) or shared (single-user mode) access token, sent
        as `Authorization: Bearer <token>`.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Unauthorized
...
    TooManyRequests:
      description: |
        Rate limit tripped by repeated failed auth attempts from this client.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Too many requests
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
    Kind:
      type: string
      description: Whether the entry is a prompt or a code snippet.
      enum: [prompt, code]
      default: prompt
    Color:
      type: string
      description: >
        Optional per-prompt color tag for visual scanning. Empty string means no
        color. Any value outside the fixed palette is coerced to "" on write.
      enum: ["", red, orange, amber, green, teal, blue, violet, pink]
      default: ""
    Template:
      type: boolean
      description: >
        Whether the entry is a reusable template offered as a starting point in
        the app's New menu. Coerced to a boolean on write.
      default: false
...
    SyncRecord:
      type: object
      description: |
        One record as it travels over the wire during sync — the full snippet
        shape keyed by `uuid` (no local `id`), plus the `deleted` tombstone flag.
        Incoming records require at least `uuid`, `title`, and `code`; anything
        else is normalized (unknown language kept, out-of-range values clamped).
      required:
        - uuid
        - title
        - description
        - code
        - language
        - tags
        - favorite
        - model
        - kind
        - copy_count
        - last_used_at
        - created_at
        - updated_at
        - deleted
      properties:
        uuid:
          type: string
          description: Stable cross-machine identity. Required.
        title:
          type: string
          maxLength: 255
        description:
          type: string
        code:
          type: string
        language:
          type: string
        tags:
          type: array
          items:
            type: string
        favorite:
          type: boolean
        model:
          type: string
        kind:
          $ref: "#/components/schemas/Kind"
        color:
          $ref: "#/components/schemas/Color"
        template:
          $ref: "#/components/schemas/Template"
        copy_count:
          type: integer
          minimum: 0
        last_used_at:
          type: [string, "null"]
        created_at:
          type: string
        updated_at:
          type: string
        deleted:
          type: boolean
          description: Tombstone flag — true means the record is deleted.